Move Rack middleware into seperate file
